requirements.txt
flask
flask-cors
flask-sqlalchemy
aiohttp
requests
python-dotenv
